Add target parameter in messages
The current implementation relies on replying to join message to
specify which user is pending challenge. Many groups prefer having
system messages deleted, and messages without a subject will cause
confusion. To mitigate spam, the user's display name will not show
if they did not pass the challenge or was removed by an admin.